requirements.txt
python-dotenv>=1.0.1
playwright>=1.54.0
pydantic>=2.8.0
PyYAML>=6.0.2
